Add support for not persisting session when maxAge is explicitly set to 0
This is especially useful in scenarios where session cookies are disabled
for some reason (e.g. not needed until authorized login). Setting maxAge
to 0 for session cookies are valid in HTTP but throws an error when used
with SETEX in redis. So setex call is avoided when maxAge is set to 0.